Add nested badge to button and restyle
