+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mechanical-wombat",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "React UI component lib for edozo apps and sites",
   "author": "edozo",
   "license": "MIT",
