--- src/ButtonWithBadge/ButtonWithBadge.styles.ts
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-import { ButtonV2 } from '../ButtonV2';
-import { Badge } from '../Badge';
-
-export const StyledButton = styled(ButtonV2)`
-  position: relative;
-`;
-
-export const StyledNewBadge = styled(Badge)`
-  position: absolute;
-  background: #2818f9;
-  color: ${p => p.theme.colors.white};
-  font-size: ${p => p.theme.spacing.xsmall};
-  letter-spacing: 0.1em;
-  font-weight: 800;
-  top: 0;
-  right: 0;
-`;
