+++ src/Header/Header.stories.tsx
@@ -8,8 +8,14 @@ import { Menu } from '../Menu';
 import { Popover } from '../Popover';
 import { List, ListItem } from '../List';
 import { ReactComponent as SwitchAppIconColour } from './switch-app-icon-color.svg';
-import { ItemWrapper, StyledButtonTitle, StyledText, SwitchAppWrapper } from '../ProductSwitch/ProductSwitch.styles';
-import { ButtonWithBadge } from '../ButtonWithBadge';
+import {
+  ItemWrapper,
+  StyledButtonTitle,
+  StyledNewBadge,
+  StyledText,
+  SwitchAppWrapper,
+} from '../ProductSwitch/ProductSwitch.styles';
+import { ButtonV2 } from '../ButtonV2';
 
 export default {
   title: 'Components/Header',
@@ -296,12 +302,15 @@ WithNewMenuPopover.args = {
               }
             >
               <div>
-                <ButtonWithBadge onClick={() => console.log('click')} badge="NEW">
+                <ButtonV2 onClick={() => console.log('click')} radius="round" variant="white">
                   <SwitchAppWrapper>
                     <SwitchAppIconColour />
                     <StyledButtonTitle>Products</StyledButtonTitle>
                   </SwitchAppWrapper>
-                </ButtonWithBadge>
+                  <StyledNewBadge color="#ffffff" background="#2818f9" size="small">
+                    NEW
+                  </StyledNewBadge>
+                </ButtonV2>
               </div>
             </Popover>
           </div>
